fix(floating-pane): guard onNodeDelete for ALL block moves, not just redock (#1662)
Follow-up to the redock guard: the same onNodeDelete -> DeleteBlock race also
fires on TEAR-OFF. Tearing a pane out moves the block to a new floater
workspace and removes the source window's layout node; that node removal fires
tabcontent's onNodeDelete -> DeleteBlock, which deletes the just-torn-off block,
so the new floater renders only the empty-tab logo (host log: object.DeleteBlock
racing workspace.TearOffBlock). It's a race, so a retry usually wins.

Generalize the renderer-local guard (redock-guard.ts -> block-move-guard.ts):
arm it at the start of any block move — redock (floater) AND pane tear-off (the
three CrossWindowDragMonitor platforms) — before the source-node removal can
fire. onNodeDelete skips DeleteBlock while it's active. Time-boxed (3-5s,
auto-expiring) so it can never permanently swallow a genuine pane close, and
released immediately when a redock resolves to no move.
