+++ frontend/app/workspace/floating-pane-workspace.tsx
@@ -41,7 +41,7 @@
 import { invokeCommand, listenEvent } from "@/app/platform/ipc";
 import { isLinux, isMacOS, isWindows } from "@/util/platformutil";
 import { FLOATER_EDGE_RESIZE_BORDER } from "@/app/workspace/floater-resize";
-import { beginRedockGuard, endRedockGuard } from "@/app/workspace/redock-guard";
+import { beginBlockMoveGuard, endBlockMoveGuard } from "@/app/workspace/block-move-guard";
 import { ErrorBoundary } from "@/app/element/errorboundary";
 import { CenteredDiv } from "@/app/element/quickelems";
 import { ModalsRenderer } from "@/app/modals/modalsrenderer";
@@ -833,13 +833,13 @@ function FloatingPaneWorkspaceElem(): JSX.Element {
             // floater's node removal triggers tabcontent's onNodeDelete, which
             // would otherwise DeleteBlock the block we're redocking (#1662).
             // Released below if no redock actually happened (drop on desktop).
-            beginRedockGuard();
+            beginBlockMoveGuard(5000);
             let redocked = false;
             try {
                 redocked = (await tryRedockAtCursorInner(screenX, screenY)) === true;
             } finally {
                 setRedockInProgress(false);
-                if (!redocked) endRedockGuard();
+                if (!redocked) endBlockMoveGuard();
             }
         };
